server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
apscheduler==3.10.4
schedule==1.2.0
supabase==2.3.0
